Ladders and elevator tilemap
ladders, doors and elevator tilemaps have been added, ill decorate the level for friday, this is just so that robin and aiden can work on the ladder scripts if they want to. hopefully we'll have a traversable level for friday. now how do we want to set up progression in it we'll have to discuss together